requirements.txt
# --------- pytorch --------- #
numpy<2
torch
torchvision
pytorch-lightning==1.9.1
torchmetrics

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
tensorboard

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests

# --------- notebook --------- #
jupyter
jupyterlab

# --------- models --------- #
segmentation-models-pytorch

# --------- data --------- #
tifffile
imagecodecs
pandas
scikit-learn

# --------- visualizations --------- #
matplotlib
Pillow
seaborn
git+https://gitlab.fbk.eu/dsip/dsip_meteo/performance-diagram